+++ CHANGELOG.md
@@ -41,6 +41,7 @@ CHANGELOG
 ### Tools
 - `intelmqctl`: `upgrade-config`: Allow setting the state file location with the `--state-file` parameter.
 - `intelmqsetup`: Add argument parsing and an option to skip setting file ownership, possibly not requiring root permissions.
+- `intelmq_generate_misp_objects_templates.py`: Tool to create a MISP object template (#1470).
 
 ### Contrib